+++ src/main/java/kmlib/starsector/markets/colonies/LocationColonies.java
@@ -1,4 +1,4 @@
-package kmlib.starsector.colonies;
+package kmlib.starsector.markets.colonies;
 
 import com.fs.starfarer.api.campaign.LocationAPI;
 import com.fs.starfarer.api.campaign.SectorAPI;