+++ src/main/java/kmlib/starsector/systems/SystemColoniesIndex.java
@@ -4,8 +4,8 @@ import com.fs.starfarer.api.campaign.SectorAPI;
 import com.fs.starfarer.api.campaign.StarSystemAPI;
 
 import kmlib.starsector.SectorWalkCounters;
-import kmlib.starsector.colonies.Colonies;
-import kmlib.starsector.colonies.SystemColonies;
+import kmlib.starsector.markets.colonies.Colonies;
+import kmlib.starsector.markets.colonies.SystemColonies;
 import kmlib.text.KmlibStrings;
 
 import java.util.Collections;