+++ src/test/java/kmlib/console/ListFactionsCommandTest.java
@@ -13,8 +13,8 @@ import com.fs.starfarer.api.characters.RelationshipAPI;
 import kmlib.console.ListFactionsCommand.FactionListingFilter;
 import kmlib.starsector.factions.FactionCustomFixture;
 import kmlib.testfixtures.console.output.CommandOutputFake;
-import kmlib.testfixtures.starsector.colonies.ColonyMarketFixture;
-import kmlib.testfixtures.starsector.colonies.ColonyPlacementFixture;
+import kmlib.testfixtures.starsector.markets.colonies.ColonyMarketFixture;
+import kmlib.testfixtures.starsector.markets.colonies.ColonyPlacementFixture;
 
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;