+++ src/test/java/kmlib/starsector/markets/colonies/HyperspaceColoniesTest.java
@@ -1,12 +1,12 @@
-package kmlib.starsector.colonies;
+package kmlib.starsector.markets.colonies;
 
 import com.fs.starfarer.api.campaign.LocationAPI;
 import com.fs.starfarer.api.campaign.SectorAPI;
 import com.fs.starfarer.api.campaign.econ.EconomyAPI;
 import com.fs.starfarer.api.campaign.econ.MarketAPI;
 
-import kmlib.testfixtures.starsector.colonies.ColonyMarketFixture;
-import kmlib.testfixtures.starsector.colonies.ColonyPlacementFixture;
+import kmlib.testfixtures.starsector.markets.colonies.ColonyMarketFixture;
+import kmlib.testfixtures.starsector.markets.colonies.ColonyPlacementFixture;
 
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;