+++ src/test/java/kmlib/starsector/markets/colonies/SystemColoniesTest.java
@@ -1,9 +1,9 @@
-package kmlib.starsector.colonies;
+package kmlib.starsector.markets.colonies;
 
 import com.fs.starfarer.api.campaign.SectorAPI;
 import com.fs.starfarer.api.campaign.StarSystemAPI;
 
-import kmlib.testfixtures.starsector.colonies.ColonyFixture;
+import kmlib.testfixtures.starsector.markets.colonies.ColonyFixture;
 
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;