+++ src/test/java/kmlib/starsector/systems/SystemColoniesIndexTest.java
@@ -4,10 +4,10 @@ import com.fs.starfarer.api.campaign.SectorAPI;
 
 import kmlib.starsector.SectorWalkCounters;
 import kmlib.starsector.WalkCountCapture;
-import kmlib.starsector.colonies.Colonies;
-import kmlib.starsector.colonies.Colony;
-import kmlib.starsector.colonies.SystemColonies;
-import kmlib.testfixtures.starsector.colonies.ColonyFixture;
+import kmlib.starsector.markets.colonies.Colonies;
+import kmlib.starsector.markets.colonies.Colony;
+import kmlib.starsector.markets.colonies.SystemColonies;
+import kmlib.testfixtures.starsector.markets.colonies.ColonyFixture;
 
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;